rename utils to configuration
